실전 매매 지원: 예수금 기반 동적 포지션 사이징, DART 재무제표 연동, KRX 종목/가격 동기화, 대시보드 제어 패널 추가

+++ src/main/java/com/canagent/service/notification/NotificationEvent.java
@@ -11,7 +11,7 @@ public record NotificationEvent(
         TradeType tradeType,
         String stockCode,
         String stockName,
-        int quantity,
+        BigDecimal quantity,
         BigDecimal price,
         BigDecimal totalAmount,
         BigDecimal profitRate,
